refactor(specification): reorganize application pieces
