+++ api/apps/api/src/modules/specification/application/__tests__/submit-specification.handler.fixtures.ts
@@ -5,10 +5,12 @@ import { CqrsModule, EventBus, IEvent } from '@nestjs/cqrs';
 import { SpecificationRepository } from '../specification.repository';
 import { SubmitSpecificationHandler } from '../submit-specification.handler';
 
-import { SubmitSpecification } from '../../commands/submit-specification.command';
-import { SpecificationOperation } from '../../feature-config';
-import { SpecificationCandidateCreated } from '../../events/specification-candidate-created.event';
-import { Specification } from '../../specification';
+import { SubmitSpecification } from '../submit-specification.command';
+import {
+  SpecificationOperation,
+  SpecificationCandidateCreated,
+  Specification,
+} from '../../domain';
 
 export const getFixtures = async () => {
   const scenarioId = v4();
