+++ api/apps/api/src/modules/specification/application/__tests__/submit-specification.handler.spec.ts
@@ -1,5 +1,5 @@
 import { PromiseType } from 'utility-types';
-import { getFixtures } from './fixtures';
+import { getFixtures } from './submit-specification.handler.fixtures';
 
 let fixtures: PromiseType<ReturnType<typeof getFixtures>>;
