+++ api/apps/api/src/modules/specification/application/determine-features.command.ts
@@ -1,0 +1,8 @@
+import { Command } from '@nestjs-architects/typed-cqrs';
+import { DetermineFeaturesInput } from '../domain';
+
+export class DetermineFeatures extends Command<void> {
+  constructor(public readonly featuresConfig: DetermineFeaturesInput) {
+    super();
+  }
+}
