+++ api/apps/api/src/modules/specification/application/specification.repository.ts
@@ -1,4 +1,4 @@
-import { Specification } from '../specification';
+import { Specification } from '../domain';
 
 export abstract class SpecificationRepository {
   abstract save(specification: Specification): Promise<void>;
