--- api/apps/api/src/modules/specification/calculate-features/application/specification.repository.ts
@@ -1,7 +1,0 @@
-import { Specification } from '../../specification';
-
-export abstract class SpecificationRepository {
-  abstract save(specification: Specification): Promise<void>;
-
-  abstract getById(id: string): Promise<Specification | undefined>;
-}
