--- api/apps/api/src/modules/specification/commands/determine-features.command.ts
@@ -1,12 +1,0 @@
-import { Command } from '@nestjs-architects/typed-cqrs';
-import { FeatureConfigInput, FeatureState } from '../feature-config';
-
-export type DeterminedFeatures = FeatureConfigInput & {
-  features: FeatureState[];
-};
-
-export class DetermineFeatures extends Command<void> {
-  constructor(public readonly featuresConfig: DeterminedFeatures) {
-    super();
-  }
-}
