+++ api/apps/api/src/modules/specification/domain/__tests__/create-candidate/scenario-specification-create.spec.ts
