+++ api/apps/api/src/modules/specification/domain/__tests__/mark-as-calculated/scenario-specification-mark-as-calculated.spec.ts
