+++ api/apps/api/src/modules/specification/domain/determine-features-input.ts
@@ -1,0 +1,5 @@
+import { FeatureConfigInput, FeatureState } from './feature-config';
+
+export type DetermineFeaturesInput = FeatureConfigInput & {
+  features: FeatureState[];
+};
