+++ api/apps/api/src/modules/specification/domain/events/specification-got-ready.event.ts
