+++ api/apps/api/src/modules/specification/domain/events/specification-published.event.ts
