+++ api/apps/api/src/modules/specification/domain/index.ts
@@ -1,0 +1,11 @@
+export * from './feature-config';
+export { Specification } from './specification';
+export {
+  SpecificationSnapshot,
+  SpecificationSnapshotInput,
+} from './specification.snapshot';
+
+export { DetermineFeaturesInput } from './determine-features-input';
+export { SpecificationGotReady } from './events/specification-got-ready.event';
+export { SpecificationPublished } from './events/specification-published.event';
+export { SpecificationCandidateCreated } from './events/specification-candidate-created.event';
