+++ api/apps/api/src/modules/specification/domain/specification.snapshot.ts
